LPD-93809 Move components to dedicated folder
