+++ modules/apps/frontend-data-set/frontend-data-set-fragment-web/src/main/resources/META-INF/resources/js/components/EntitySelectorRow.tsx
@@ -9,7 +9,7 @@ import {openSelectionModal} from 'frontend-js-components-web';
 import {fetch, sub} from 'frontend-js-web';
 import React from 'react';
 
-import {IContentMappedTokenValue} from './tokenMapping';
+import {IContentMappedTokenValue} from '../tokenMapping';
 
 const EDITOR_PORTLET_ID =
 	'com_liferay_layout_content_page_editor_web_internal_portlet_ContentPageEditorPortlet';
